+++ server/auth.js
@@ -3,6 +3,11 @@
 import type { QConfig } from "./config";
 import fetch from 'node-fetch';
 
+export type AccessKey = {
+    key: string,
+    restrictToAccounts?: string[],
+}
+
 export type AccessRights = {
     granted: bool,
     restrictToAccounts: string[],
@@ -18,15 +23,16 @@ const deniedAccess: AccessRights = Object.freeze({
     restrictToAccounts: [],
 });
 
-export default class QAuth {
+export class Auth {
     config: QConfig;
 
     constructor(config: QConfig) {
         this.config = config;
     }
 
-    static extractAccessKey(req: any): string {
-        return (req && req.headers && req.headers.authorization) || ''
+    static extractAccessKey(req: any, connection: any): string {
+        return (req && req.headers && (req.headers.accessKey || req.headers.accesskey))
+            || (connection && connection.context && connection.context.accessKey);
     }
 
     static error(code: number, message: string): Error {
@@ -36,19 +42,20 @@ export default class QAuth {
         return error;
     }
 
+    static unauthorizedError(): Error {
+        return Auth.error(401, 'Unauthorized');
+    }
+
     authServiceRequired() {
         if (!this.config.authorization.endpoint) {
-            throw QAuth.error(500, 'Auth service unavailable');
+            throw Auth.error(500, 'Auth service unavailable');
         }
     }
 
     async requireGrantedAccess(accessKey: string | typeof undefined): Promise<AccessRights> {
         const access = await this.getAccessRights(accessKey);
         if (!access.granted) {
-            throw QAuth.error(401, 'Unauthorized');
-        }
-        if (access.restrictToAccounts.length > 0) {
-            throw QAuth.error(500, 'Internal error: GraphQL services doesn\'t support account restrictions yet');
+            throw Auth.unauthorizedError();
         }
         return access;
     }
@@ -60,9 +67,13 @@ export default class QAuth {
         if ((accessKey || '') === '') {
             return deniedAccess;
         }
-        return this.invokeAuth('getAccessRights', {
+        const rights = await this.invokeAuth('getAccessRights', {
             accessKey,
         });
+        if (!rights.restrictToAccounts) {
+            rights.restrictToAccounts = [];
+        }
+        return rights;
     }
 
     async getManagementAccessKey(): Promise<string> {
@@ -72,7 +83,7 @@ export default class QAuth {
 
     async registerAccessKeys(
         account: string,
-        keys: string[],
+        keys: AccessKey[],
         signedManagementAccessKey: string
     ): Promise<number> {
         this.authServiceRequired();