+++ server/server.js
@@ -19,8 +19,10 @@ import fs from 'fs';
 import express from 'express';
 import http from 'http';
 
-import { ApolloServer } from 'apollo-server-express';
-
+import { ApolloServer, ApolloServerExpressConfig } from 'apollo-server-express';
+import { ConnectionContext } from 'subscriptions-transport-ws';
+import type { TONClient } from "ton-client-js/types";
+import { TONClient as TONClientNodeJs } from 'ton-client-node-js';
 import Arango from './arango';
 
 import { createResolvers } from './resolvers-generated';
@@ -32,7 +34,7 @@ import QLogs from './logs';
 import type { QLog } from './logs';
 import { QTracer } from "./tracer";
 import { Tracer } from "opentracing";
-import QAuth from './q-auth';
+import { Auth } from './auth';
 
 type QOptions = {
     config: QConfig,
@@ -55,7 +57,8 @@ export default class TONQServer {
     endPoints: EndPoint[];
     db: Arango;
     tracer: Tracer;
-    auth: QAuth;
+    client: TONClient;
+    auth: Auth;
     shared: Map<string, any>;
 
 
@@ -65,11 +68,11 @@ export default class TONQServer {
         this.log = this.logs.create('server');
         this.shared = new Map();
         this.tracer = QTracer.create(options.config);
-        this.auth = new QAuth(options.config);
+        this.auth = new Auth(options.config);
         this.endPoints = [];
         this.app = express();
         this.server = http.createServer(this.app);
-        this.db = new Arango(this.config, this.logs, this.tracer);
+        this.db = new Arango(this.config, this.logs, this.auth, this.tracer);
         this.addEndPoint({
             path: '/graphql/mam',
             resolvers: resolversMam,
@@ -86,6 +89,7 @@ export default class TONQServer {
 
 
     async start() {
+        this.client = await TONClientNodeJs.create({ servers: [''] });
         await this.db.start();
         const { host, port } = this.config.server;
         this.server.listen({ host, port }, () => {
@@ -100,22 +104,31 @@ export default class TONQServer {
         const typeDefs = endPoint.typeDefFileNames
             .map(x => fs.readFileSync(x, 'utf-8'))
             .join('\n');
-        const apollo = new ApolloServer({
+        const config: ApolloServerExpressConfig = {
             typeDefs,
             resolvers: endPoint.resolvers,
+            subscriptions: {
+                onConnect(connectionParams: Object, _websocket: WebSocket, _context: ConnectionContext): any {
+                    return {
+                        accessKey: connectionParams.accessKey || connectionParams.accesskey,
+                    }
+                }
+            },
             context: ({ req, connection }) => {
                 return {
                     db: this.db,
                     tracer: this.tracer,
                     auth: this.auth,
+                    client: this.client,
                     config: this.config,
                     shared: this.shared,
                     remoteAddress: (req && req.socket && req.socket.remoteAddress) || '',
-                    accessKey: QAuth.extractAccessKey(req),
+                    accessKey: Auth.extractAccessKey(req, connection),
                     parentSpan: QTracer.extractParentSpan(this.tracer, connection ? connection : req),
                 };
             },
-        });
+        };
+        const apollo = new ApolloServer(config);
         apollo.applyMiddleware({ app: this.app, path: endPoint.path });
         if (endPoint.supportSubscriptions) {
             apollo.installSubscriptionHandlers(this.server);